Fixed Parser Bug that caused 2x Duplicate Data. Also, Optimized Parser performance.
The parser code had a subtle bug. Everytime data is received on the serial port the parser is called and the current device data is written to display (or file). The issue is that sometimes the serial port contains a partial packet. The parser is called on this data update. Then regardless of whether we parsed a (full) new packet or not we write a new row of data to screen or file.

The usb basestation adds RSSI and header making the packet size 65, but the station can only send 64 bit packets. so each umyo packet results in 2 basestation packets (64/1). At the python level this results in double the data production with ~half of them duplicates!!!!!!!

I also simplified and optimized the parser logic so it more precisely identifies, validates and processes each packet.

finally i added an illustrative simple_data_logger.py (based on turfptax's script) that logs data using the new parser approach.